Create route in prep for OpenAI API
