+++ frontend/src/components/ResumeUpload.tsx
@@ -1,5 +1,6 @@
 "use client";
 
+import { useState } from "react";
 import type { ResumeUploadState } from "@/types/resume";
 
 interface ResumeUploadProps {
@@ -7,6 +8,7 @@ interface ResumeUploadProps {
   onFileUpload: (file: File) => void;
   onFileError: (error: string) => void;
   onRemoveFile: () => void;
+  onAnalysisComplete?: (data: any) => void; // 👈 callback for API results
 }
 
 export default function ResumeUpload({
@@ -14,15 +16,17 @@ export default function ResumeUpload({
   onFileUpload,
   onFileError,
   onRemoveFile,
+  onAnalysisComplete,
 }: ResumeUploadProps) {
-  const handleFileChange = (e: React.ChangeEvent<HTMLInputElement>) => {
+  const [loading, setLoading] = useState(false);
+
+  const handleFileChange = async (e: React.ChangeEvent<HTMLInputElement>) => {
     const file = e.target.files?.[0];
     if (!file) return;
 
     if (file.size > 5 * 1024 * 1024) {
       onFileError("File size exceeds 5MB");
       return;
-      ``;
     }
 
     const allowedTypes = [
@@ -35,7 +39,33 @@ export default function ResumeUpload({
       return;
     }
 
+    // ✅ Update UI state
     onFileUpload(file);
+
+    // ✅ Send file to API
+    try {
+      setLoading(true);
+      const formData = new FormData();
+      formData.append("resume", file);
+
+      const res = await fetch("/api/analyze", {
+        method: "POST",
+        body: formData,
+      });
+
+      if (!res.ok) throw new Error("Failed to analyze resume");
+
+      const data = await res.json();
+
+      // 👇 Call parent with API results
+      if (onAnalysisComplete) {
+        onAnalysisComplete(data);
+      }
+    } catch (err: any) {
+      onFileError(err.message || "Something went wrong");
+    } finally {
+      setLoading(false);
+    }
   };
 
   return (
@@ -45,12 +75,16 @@ export default function ResumeUpload({
           <p className="text-sm text-gray-700 dark:text-gray-300">
             Uploaded: <span className="font-medium">{fileState.file.name}</span>
           </p>
-          <button
-            onClick={onRemoveFile}
-            className="px-4 py-2 text-sm rounded-full bg-red-100 text-red-600 hover:bg-red-200 dark:bg-red-900/30 dark:text-red-300 dark:hover:bg-red-900/50 transition"
-          >
-            Remove File
-          </button>
+          {loading ? (
+            <p className="text-sm text-blue-500">Analyzing resume...</p>
+          ) : (
+            <button
+              onClick={onRemoveFile}
+              className="px-4 py-2 text-sm rounded-full bg-red-100 text-red-600 hover:bg-red-200 dark:bg-red-900/30 dark:text-red-300 dark:hover:bg-red-900/50 transition"
+            >
+              Remove File
+            </button>
+          )}
         </div>
       ) : (
         <label className="w-full cursor-pointer border-2 border-dashed border-gray-300 dark:border-gray-700 rounded-2xl p-8 flex flex-col items-center justify-center bg-white/70 dark:bg-gray-800/70 backdrop-blur-sm hover:border-blue-500 hover:bg-blue-50/30 dark:hover:bg-gray-700/50 transition shadow-sm">
